# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: analytics-agent v2
  description: Contract-first API. Figures carry provenance; chart series are computed by the backend.
    501 = spec'd, not live yet.
  version: 0.7.0
paths:
...
  /sessions:
    post:
      tags:
      - sessions
      summary: Create Session
      operationId: create_session_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/SessionCreate'
              - type: 'null'
              title: Body
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCreated'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{sid}:
    get:
      tags:
      - sessions
      summary: Get Session
      operationId: get_session_sessions__sid__get
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          title: Sid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - sessions
      summary: Delete Session
      operationId: delete_session_sessions__sid__delete
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          title: Sid
      responses:
        '204':
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{sid}/ui-state:
    put:
      tags:
      - sessions
      summary: Put Ui State
      operationId: put_ui_state_sessions__sid__ui_state_put
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          title: Sid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UiStatePut'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionConflict'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /turns:
    post:
      tags:
      - turns
      summary: Create Turn
      operationId: create_turn_turns_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TurnCreate'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnCreated'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /turns/{turn_id}:
    get:
      tags:
      - turns
      summary: Get Turn
      operationId: get_turn_turns__turn_id__get
      parameters:
      - name: turn_id
        in: path
        required: true
        schema:
          type: string
          title: Turn Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Turn'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /sessions/{sid}/turns:
    get:
      tags:
      - turns
      summary: List Turns
      operationId: list_turns_sessions__sid__turns_get
      parameters:
      - name: sid
        in: path
        required: true
        schema:
          type: string
          title: Sid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnList'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{ws}/uploads:
    post:
      tags:
      - datasets
      summary: Upload
      operationId: upload_workspaces__ws__uploads_post
      parameters:
      - name: ws
        in: path
        required: true
        schema:
          type: string
          title: Ws
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_workspaces__ws__uploads_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}:
    get:
      tags:
      - datasets
      summary: Get Dataset
      operationId: get_dataset_datasets__dataset_id__get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/profile:
    get:
      tags:
      - datasets
      summary: Profile
      operationId: profile_datasets__dataset_id__profile_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/cleaning/proposals:
    get:
      tags:
      - cleaning
      summary: Cleaning Proposals
      operationId: cleaning_proposals_datasets__dataset_id__cleaning_proposals_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleaningProposals'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/cleaning/approve:
    post:
      tags:
      - cleaning
      summary: Cleaning Approve
      operationId: cleaning_approve_datasets__dataset_id__cleaning_approve_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveActions'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResult'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/domains/detect:
    get:
      tags:
      - domains
      summary: Detect
      operationId: detect_datasets__dataset_id__domains_detect_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainDetection'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/domains/confirm:
    post:
      tags:
      - domains
      summary: Confirm Domains
      operationId: confirm_domains_datasets__dataset_id__domains_confirm_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainConfirm'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmed'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/contract/proposal:
    get:
      tags:
      - contract
      summary: Contract Proposal
      operationId: contract_proposal_datasets__dataset_id__contract_proposal_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractProposal'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/contract/confirm:
    post:
      tags:
      - contract
      summary: Contract Confirm
      operationId: contract_confirm_datasets__dataset_id__contract_confirm_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractConfirm'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmed'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/tools:
    get:
      tags:
      - tools
      summary: Tools
      operationId: tools_datasets__dataset_id__tools_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolList'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/forks:
    post:
      tags:
      - contract
      summary: Answer Forks
      operationId: answer_forks_datasets__dataset_id__forks_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForkAnswers'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmed'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
...
  /datasets/{dataset_id}/validity-rules:
    get:
      tags:
      - rules
      summary: Validity Rules
      operationId: validity_rules_datasets__dataset_id__validity_rules_get
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidityRules'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
  /datasets/{dataset_id}/validity-rules/approve:
    post:
      tags:
      - rules
      summary: Approve Rules
      operationId: approve_rules_datasets__dataset_id__validity_rules_approve_post
      parameters:
      - name: dataset_id
        in: path
        required: true
        schema:
          type: string
          title: Dataset Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleApprove'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Confirmed'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Not Found
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Request Entity Too Large
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unprocessable Entity
...
components:
  schemas:
...
    ErrorBody:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - error
      title: ErrorBody
      examples:
      - error:
          code: not_found
          message: session x not found or expired
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SessionCreate:
      properties:
        workspace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Workspace Id
          description: 'Reuse a workspace (its datasets persist). Omitted: a new workspace is created.'
      type: object
      title: SessionCreate
      examples:
      - {}
      - workspace_id: ws_3f9a1c2b7d10
    SessionCreated:
      properties:
        sid:
          type: string
          title: Sid
          description: UUID4
        version:
          type: integer
          title: Version
      type: object
      required:
      - sid
      - version
      title: SessionCreated
      examples:
      - sid: 0b8e0c64-5b0e-4c55-9c1e-2f1c2b9d7a11
        version: 1
...
    TurnCreated:
      properties:
        turn_id:
          type: string
          title: Turn Id
      type: object
      required:
      - turn_id
      title: TurnCreated
      examples:
      - turn_id: t_7c1d9e0a4b2f4e59
...
    UiStatePut:
      properties:
        ui_state:
          additionalProperties: true
          type: object
          title: Ui State
          description: '<=256 KB JSON. No data rows, API keys or PII: rejected with 422'
        version:
          type: integer
          title: Version
          description: The version you last read
      type: object
      required:
      - ui_state
      - version
      title: UiStatePut
      examples:
      - ui_state:
          dataset_id: ds_ads_2026q3
          screen: contract
        version: 3
...